Adding happiness and money to customer interaction
- When customer receives their item, we add up the cost to money
- When customer leaves, we check the percent of products we delivered correctly and calculate happiness

+++ Assets/_Data/Products/Scripts/Product.cs
@@ -6,7 +6,7 @@ namespace _Data.Products
     public class Product : ScriptableObject
     {
         public string productName;
-        public float sellingPrice;
+        public int sellingPrice;
         public Sprite sprite;
         public GameObject prefab;
     }
